+++ src/lib/items/Items.js
@@ -26,7 +26,6 @@ export default class Items extends Component {
     canvasTimeStart: PropTypes.number.isRequired,
     canvasTimeEnd: PropTypes.number.isRequired,
     canvasWidth: PropTypes.number.isRequired,
-    minimumWidthForItemContentVisibility: PropTypes.number.isRequired,
 
     dragSnap: PropTypes.number,
     minResizeWidth: PropTypes.number,
@@ -79,9 +78,7 @@ export default class Items extends Component {
       nextProps.canResize === this.props.canResize &&
       nextProps.canSelect === this.props.canSelect &&
       nextProps.dimensionItems === this.props.dimensionItems &&
-      nextProps.topOffset === this.props.topOffset &&
-      nextProps.minimumWidthForItemContentVisibility ===
-        this.props.minimumWidthForItemContentVisibility
+      nextProps.topOffset === this.props.topOffset
     )
   }
 
@@ -111,7 +108,6 @@ export default class Items extends Component {
       canvasTimeStart,
       canvasTimeEnd,
       dimensionItems,
-      minimumWidthForItemContentVisibility
     } = this.props
     const { itemIdKey, itemGroupKey } = this.props.keys
 
@@ -171,9 +167,6 @@ export default class Items extends Component {
               onContextMenu={this.props.onItemContextMenu}
               onSelect={this.props.itemSelect}
               itemRenderer={this.props.itemRenderer}
-              minimumWidthForItemContentVisibility={
-                minimumWidthForItemContentVisibility
-              }
             />
           ))}
       </div>