+++ src/lib/items/defaultItemRendere.js
@@ -1,0 +1,25 @@
+import React from 'react'
+
+export const defaultItemRenderer = ({
+  item,
+  timelineContext,
+  itemContext,
+  getItemProps,
+  getResizeProps
+}) => {
+  const { left: leftResizeProps, right: rightResizeProps } = getResizeProps()
+  return (
+    <div {...getItemProps(item.itemProps)}>
+      {itemContext.useResizeHandle ? <div {...leftResizeProps} /> : ''}
+
+      <div
+        className="rct-item-content"
+        style={{ maxHeight: `${itemContext.dimensions.height}` }}
+      >
+        {itemContext.title}
+      </div>
+
+      {itemContext.useResizeHandle ? <div {...rightResizeProps} /> : ''}
+    </div>
+  )
+}